refactor: variant dialog
