+++ src/component/feature/variant/AddVariant/AddVariant.styles.js
@@ -1,0 +1,3 @@
+import { makeStyles } from '@material-ui/styles';
+
+export const useStyles = makeStyles(theme => ({}));
