+++ src/component/feature/variant/AddVariant/OverrideConfig/OverrideConfig.styles.js
@@ -1,0 +1,7 @@
+import { makeStyles } from '@material-ui/styles';
+
+export const useStyles = makeStyles(theme => ({
+    contextFieldSelect: {
+        marginRight: '8px',
+    },
+}));
